Move enum method from helper to enum helper
