+++ corekit/src/main/java/com/midtrans/sdk/corekit/utilities/Helper.java
@@ -77,80 +77,6 @@ public class Helper {
         }
     }
 
-    public static Map<String, List<Integer>> mappingMapToBankTypeForSet(Map<String, List<Integer>> installmentMap) {
-        HashMap<String, List<Integer>> mappedHash = new HashMap<>();
-        for (Map.Entry<String, List<Integer>> entry : installmentMap.entrySet()) {
-            mappedHash.put(mappingToBankType(entry.getKey()), entry.getValue());
-        }
-        return mappedHash;
-    }
-
-    public static @ExpiryTimeUnit
-    String mappingToExpiryTimeUnit(@ExpiryTimeUnit String unit) {
-        switch (unit) {
-            case ExpiryTimeUnit.DAY:
-                return ExpiryTimeUnit.DAY;
-            case ExpiryTimeUnit.HOUR:
-                return ExpiryTimeUnit.HOUR;
-            case ExpiryTimeUnit.MINUTE:
-                return ExpiryTimeUnit.MINUTE;
-        }
-        return null;
-    }
-
-    public static @Authentication
-    String mappingToCreditCardAuthentication(@Authentication String authentication, boolean secure) {
-        if (authentication.equalsIgnoreCase(Authentication.AUTH_3DS) && secure) {
-            return Authentication.AUTH_3DS;
-        } else if (authentication.equalsIgnoreCase(Authentication.AUTH_RBA) && !secure) {
-            return Authentication.AUTH_RBA;
-        } else {
-            return Authentication.AUTH_NONE;
-        }
-    }
-
-    public static @AcquiringChannel
-    String mappingToAcquiringChannel(@AcquiringChannel String acquiringChannel) {
-        if (acquiringChannel.equalsIgnoreCase(AcquiringChannel.MIGS) || acquiringChannel == null) {
-            return AcquiringChannel.MIGS;
-        } else {
-            return null;
-        }
-    }
-
-    public static @CreditCardTransactionType
-    String mappingToCreditCardType(@CreditCardTransactionType String type) {
-        if (type.equalsIgnoreCase(CreditCardTransactionType.AUTHORIZE)) {
-            return CreditCardTransactionType.AUTHORIZE;
-        } else {
-            return CreditCardTransactionType.AUTHORIZE_CAPTURE;
-        }
-    }
-
-    public static String mappingToBankType(@AcquiringBankType String bank) {
-        if (bank.equals(AcquiringBankType.CIMB)) {
-            return AcquiringBankType.CIMB;
-        } else if (bank.equals(AcquiringBankType.BCA)) {
-            return AcquiringBankType.BCA;
-        } else if (bank.equals(AcquiringBankType.MANDIRI)) {
-            return AcquiringBankType.MANDIRI;
-        } else if (bank.equals(AcquiringBankType.BNI)) {
-            return AcquiringBankType.BNI;
-        } else if (bank.equals(AcquiringBankType.BRI)) {
-            return AcquiringBankType.BRI;
-        } else if (bank.equals(AcquiringBankType.DANAMON)) {
-            return AcquiringBankType.DANAMON;
-        } else if (bank.equals(AcquiringBankType.MAYBANK)) {
-            return AcquiringBankType.MAYBANK;
-        } else if (bank.equals(AcquiringBankType.MEGA)) {
-            return AcquiringBankType.MEGA;
-        } else if (bank.equals(AcquiringBankType.NONE)) {
-            return AcquiringBankType.NONE;
-        } else {
-            return null;
-        }
-    }
-
     public static String getDeviceType(Activity activity) {
         String deviceType;
         DisplayMetrics metrics = new DisplayMetrics();
